requirements.txt
google-auth-oauthlib==0.7.1
google-api-python-client==2.84.0
python-dotenv==1.0.0
openai==1.63.0
